+++ python/pathway/xpacks/llm/llms.py
@@ -699,6 +699,6 @@ def prompt_chat_single_qa(question: str) -> pw.Json:
     >>> r = t.select(prompt=llms.prompt_chat_single_qa(t.txt))
     >>> pw.debug.compute_and_print(r, include_id=False)
     prompt
-    [{"content": "Wazzup?", "role": "system"}]
+    [{"role": "system", "content": "Wazzup?"}]
     """
     return pw.Json([dict(role="system", content=question)])